+++ example/grizzly_primitives_example.dart
@@ -5,7 +5,7 @@ main() {
   print(p1.key); // => 1
   print(p1.value); // => one
 
-  final i1 = idx1D(5);
+  final Index1D i1 = idx1D(5);
   print(i1.dim); // => 1
   print(i1.x); // => 5
   print(i1[0]); // => 5