+++ lib/src/core/core.dart
@@ -1,0 +1,7 @@
+library grizzly.primitives;
+
+import 'package:meta/meta.dart';
+
+part 'package:grizzly_primitives/src/core/pair.dart';
+part 'package:grizzly_primitives/src/core/extent.dart';
+part 'package:grizzly_primitives/src/core/index.dart';