+++ lib/src/src.dart
@@ -1,7 +1,0 @@
-library grizzly.primitives;
-
-import 'package:meta/meta.dart';
-
-part 'pair.dart';
-part 'extent.dart';
-part 'index.dart';