+++ CHANGELOG.md
@@ -1,5 +1,34 @@
 # Changelog
 
+## 0.0.9
+
++ Removed `uniqueSqueeze`
++ Added `sort` to Array and Array2D
+
+## 0.0.8
+
++ Added `unique` and `uniqueSqueeze` to Array and Array2D
+
+## 0.0.7
+
++ `Array` and `Array2D` do not inherit from `Iterable`
+
+## 0.0.5
+
++ Removed `valueCounts` from `Array` and `Array2D`
+
+## 0.0.4
+
++ `Array`
++ `Numeric1D`
++ `Array2D`
++ `Numeric2D`
+
+## 0.0.3
+
++ `Series`
++ `DataFrameBase`
+
 ## 0.0.2
 
 - `Extent` gets custom `==` operator